debug and verif NC

+++ data_export/results_export/DRIAS_1D/3_les_dimensions.R
@@ -94,7 +94,7 @@ to = as.POSIXct(max(Date), tz=fuseau_horaire)
 origin = as.POSIXct("1950-01-01", tz=fuseau_horaire)
 units = paste0("days since ", origin)
 time = seq.POSIXt(from=from, to=to, by=timestep)
-time = as.numeric(time - origin)
+time = as.integer(time - origin)
 
 NCf$time.name = "time"
 NCf$time.value = time
